+++ radio/src/targets/horus/CMakeLists.txt
@@ -288,13 +288,13 @@ if (TELEMETRY_MAVLINK)
     add_definitions(-DTELEMETRY_MAVLINK_USB_SERIAL)
   endif()
   set(SRC ${SRC} 
-    mavlink_telem.cpp
-    mavlink_telem_interface.cpp
-    mavlink_telem_mavapi.cpp
-    mavlink_telem_autopilot.cpp 
-    mavlink_telem_camera.cpp
-    mavlink_telem_gimbal.cpp 
-    mavlink_telem_qshot.cpp
+    telemetry/mavlink/mavlink_telem.cpp
+    telemetry/mavlink/mavlink_telem_interface.cpp
+    telemetry/mavlink/mavlink_telem_mavapi.cpp
+    telemetry/mavlink/mavlink_telem_autopilot.cpp 
+    telemetry/mavlink/mavlink_telem_camera.cpp
+    telemetry/mavlink/mavlink_telem_gimbal.cpp 
+    telemetry/mavlink/mavlink_telem_qshot.cpp
     )
   if(NOT LUA STREQUAL NO)
     set(SRC ${SRC} lua/api_mavlink.cpp lua/api_mavsdk.cpp)
