+++ radio/src/telemetry/mavlink/mavlink_telem_gimbal.cpp
